feat: refactoring folder names according to convention
